SSL certificate details
This script can be used to export SSL certificate details from internal as well as external web servers. The output is exported to a CSV file. The script is helpful when it comes to quickly gather certificate from large number of servers.

This script calls another script (Get-RemoteSSLCerts.ps1) from within. Place both the parent and child scripts in the same directory and execute it.