List unused/obsolete/inactive AD user accounts
This script can be used to automate the reporting of unused/obsolete/inactive AD user account which have not logged in since specific number of days. 

Several OUs like Service Accounts, Disabled Users, etc. can be filtered out from the search. The exceptions can be defined under # Defining Exceptions section in the script.

Email report in HTML format is sent to the predefined recipients and this information can be used to move the AD user accounts to the Disabled Users OU, if required.